requirements.txt
asgiref==3.4.1
autopep8==1.6.0
Django==3.2.7
django-ckeditor==6.2.0
django-debug-toolbar==3.2.4
django-js-asset==1.2.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
mysqlclient==2.1.0
Pillow==8.3.2
pycodestyle==2.8.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
